libvncclient: make client callback indicative for Extended Clipboard
A client which has GotXCutTextUTF8 set indicates it can handle UTF8
cuttext from the server. Older clients will not have this set and
thus not opt in for UTF8 (and depending on the server implementation,
thus not opt out from Latin1, resulting in no cuttext from server at
all for them).